PR feedback - renamed project to -swagger-ui
