+++ contentgrid-spring-swagger-ui/build.gradle
