+++ contentgrid-spring-swagger-ui/src/main/resources/swagger-initializer.js
