+++ settings.gradle
@@ -26,4 +26,4 @@ include 'contentgrid-spring-boot-platform'
 include 'contentgrid-spring-data-rest'
 include 'contentgrid-spring-integration-events'
 include 'contentgrid-spring-querydsl'
-include 'contentgrid-spring-swagger'
+include 'contentgrid-spring-swagger-ui'
